+++ anyline-tiretread-react-native/ios/AnylineTtrMobileWrapperReactNative.mm
@@ -1,70 +1,102 @@
-#ifdef RCT_NEW_ARCH_ENABLED
-#import <AnylineTtrMobileWrapperReactNativeSpec/AnylineTtrMobileWrapperReactNativeSpec.h>
-#endif
+#import "AnylineTtrMobileWrapperReactNative.h"
+#import "AnylineTtrPlugin.h"
 
-#import <React/RCTBridgeModule.h>
+@interface AnylineTtrMobileWrapperReactNative ()
 
-@interface RCT_EXTERN_MODULE(AnylineTtrMobileWrapperReactNative, NSObject)
+@property(nonatomic, strong) AnylineTtrPlugin *plugin;
 
-RCT_EXTERN_METHOD(initialize:(NSDictionary)options
-                 withResolver:(RCTPromiseResolveBlock)resolve
-                 withRejecter:(RCTPromiseRejectBlock)reject)
+@end
 
-RCT_EXTERN_METHOD(scan:(NSDictionary)options
-                 withResolver:(RCTPromiseResolveBlock)resolve
-                 withRejecter:(RCTPromiseRejectBlock)reject)
+@implementation AnylineTtrMobileWrapperReactNative
 
-RCT_EXTERN_METHOD(getResult:(NSDictionary)options
-                 withResolver:(RCTPromiseResolveBlock)resolve
-                 withRejecter:(RCTPromiseRejectBlock)reject)
+RCT_EXPORT_MODULE()
 
-RCT_EXTERN_METHOD(getHeatmap:(NSDictionary)options
-                 withResolver:(RCTPromiseResolveBlock)resolve
-                 withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(setTestingConfig:(NSDictionary)options
-                 withResolver:(RCTPromiseResolveBlock)resolve
-                 withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(clearTestingConfig:(RCTPromiseResolveBlock)resolve
-                 withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(sendCommentFeedback:(NSDictionary)options
-                 withResolver:(RCTPromiseResolveBlock)resolve
-                 withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(sendTreadDepthResultFeedback:(NSDictionary)options
-                 withResolver:(RCTPromiseResolveBlock)resolve
-                 withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(sendTireIdFeedback:(NSDictionary)options
-                 withResolver:(RCTPromiseResolveBlock)resolve
-                 withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(getSdkVersion:(RCTPromiseResolveBlock)resolve
-                 withRejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(getWrapperVersion:(RCTPromiseResolveBlock)resolve
-                 withRejecter:(RCTPromiseRejectBlock)reject)
-
-+ (BOOL)requiresMainQueueSetup
+- (instancetype)init
 {
-  return NO;
+  self = [super init];
+  if (self) {
+    _plugin = [AnylineTtrPlugin new];
+  }
+  return self;
+}
+
+- (dispatch_queue_t)methodQueue
+{
+  return dispatch_get_main_queue();
+}
+
+RCT_EXPORT_METHOD(initialize:(NSDictionary *)options
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  [self.plugin initializeWithOptions:options resolver:resolve rejecter:reject];
+}
+
+RCT_EXPORT_METHOD(scan:(NSDictionary *)options
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  [self.plugin scanWithOptions:options resolver:resolve rejecter:reject];
+}
+
+RCT_EXPORT_METHOD(getResult:(NSDictionary *)options
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  [self.plugin getResultWithOptions:options resolver:resolve rejecter:reject];
+}
+
+RCT_EXPORT_METHOD(getHeatmap:(NSDictionary *)options
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  [self.plugin getHeatmapWithOptions:options resolver:resolve rejecter:reject];
+}
+
+RCT_EXPORT_METHOD(setTestingConfig:(NSDictionary *)options
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  [self.plugin setTestingConfigWithOptions:options resolver:resolve rejecter:reject];
+}
+
+RCT_EXPORT_METHOD(clearTestingConfig:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  [self.plugin clearTestingConfigWithResolver:resolve rejecter:reject];
+}
+
+RCT_EXPORT_METHOD(sendCommentFeedback:(NSDictionary *)options
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  [self.plugin sendCommentFeedbackWithOptions:options resolver:resolve rejecter:reject];
+}
+
+RCT_EXPORT_METHOD(sendTreadDepthResultFeedback:(NSDictionary *)options
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  [self.plugin sendTreadDepthResultFeedbackWithOptions:options resolver:resolve rejecter:reject];
+}
+
+RCT_EXPORT_METHOD(sendTireIdFeedback:(NSDictionary *)options
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  [self.plugin sendTireIdFeedbackWithOptions:options resolver:resolve rejecter:reject];
+}
+
+RCT_EXPORT_METHOD(getSdkVersion:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  [self.plugin getSdkVersionWithResolver:resolve rejecter:reject];
+}
+
+RCT_EXPORT_METHOD(getWrapperVersion:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  [self.plugin getWrapperVersionWithResolver:resolve rejecter:reject];
 }
 
 @end
-
-#ifdef RCT_NEW_ARCH_ENABLED
-@interface AnylineTtrMobileWrapperReactNative (NativeAnylineTtrModuleSpecConformance) <NativeAnylineTtrModuleSpec>
-@end
-
-@implementation AnylineTtrMobileWrapperReactNative (TurboModule)
-
-- (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:
-    (const facebook::react::ObjCTurboModule::InitParams &)params
-{
-  return std::make_shared<facebook::react::NativeAnylineTtrModuleSpecJSI>(params);
-}
-
-@end
-#endif